Rename 'NG Playlist' directory ng_playlist
