+++ foo_ui_columns/callbacks.cpp
@@ -1,5 +1,5 @@
 #include "stdafx.h"
-#include "NG Playlist/ng_playlist.h"
+#include "ng_playlist/ng_playlist.h"
 #include "seekbar.h"
 #include "playlist_view.h"
 #include "main_window.h"
