+++ foo_ui_columns/config_artwork.cpp
@@ -1,5 +1,5 @@
 #include "stdafx.h"
-#include "NG Playlist/ng_playlist.h"
+#include "ng_playlist/ng_playlist.h"
 #include "artwork.h"
 #include "config.h"
