+++ foo_ui_columns/config_columns_v2.cpp
@@ -1,5 +1,5 @@
 #include "stdafx.h"
-#include "NG Playlist/ng_playlist.h"
+#include "ng_playlist/ng_playlist.h"
 #include "config.h"
 #include "config_columns_v2.h"
