+++ foo_ui_columns/ng_playlist/ng_playlist_callbacks.cpp
