+++ foo_ui_columns/playlist_view_msgproc.cpp
@@ -1,5 +1,5 @@
 #include "stdafx.h"
-#include "NG Playlist/ng_playlist.h"
+#include "ng_playlist/ng_playlist.h"
 #include "playlist_view.h"
 #include "font_notify.h"
 #include "config_columns_v2.h"
